Switch to edomora97 official repo and fix a import bug

+++ rust/turingarena-contest/src/main.rs
@@ -99,7 +99,7 @@ fn main() -> Result<()> {
             format,
             force,
         } => {
-            if force {
+            if force && context.database_url.exists() {
                 std::fs::remove_file(&context.database_url)?;
             }
             formats::import(&context, &path, &format)
